+++ src/ThunderEgg/FineNbrInfo.h
@@ -32,7 +32,7 @@ namespace ThunderEgg
  *
  * @tparam D the number of Cartesian dimensions.
  */
-template <size_t D> class FineNbrInfo : public NbrInfo<D>
+template <int D> class FineNbrInfo : public NbrInfo<D>
 {
 	public:
 	/**
