+++ src/ThunderEgg/PatchSolver.h
@@ -36,7 +36,7 @@ namespace ThunderEgg
  *
  * @tparam D the number of cartesian dimensions
  */
-template <size_t D> class PatchSolver : public virtual Operator<D>, public virtual GMG::Smoother<D>
+template <int D> class PatchSolver : public virtual Operator<D>, public virtual GMG::Smoother<D>
 {
 	protected:
 	/**
