+++ src/ThunderEgg/Poisson/DFTPatchSolver.h
@@ -41,7 +41,7 @@ namespace Poisson
  *
  * @tparam D the number of Cartesian dimensions
  */
-template <size_t D> class DFTPatchSolver : public PatchSolver<D>
+template <int D> class DFTPatchSolver : public PatchSolver<D>
 {
 	private:
 	/**
